requirements.txt
python-telegram-bot==22.5
nest_asyncio
httpx
